+++ Makefile
@@ -2,7 +2,7 @@ GIT_COMMIT=$(shell git rev-parse HEAD)
 LDFLAGS=-ldflags "-X main.GitCommit=$(GIT_COMMIT)"
 
 build:
-	@go build $(LDFLAGS) -o bin/slip slip/main.go
+	@go build $(LDFLAGS) -o bin/slip main.go
 
 test:
 	@go test ./...
