+++ internal/slip.go
@@ -1,4 +1,4 @@
-package slip
+package internal
 
 import (
 	"fmt"
@@ -8,9 +8,6 @@ import (
 
 	"github.com/peterh/liner"
 )
-
-////////////////////////////////////////////////////////////////////////////////
-// Slip
 
 type Slip struct {
 	env    *Enviroment
@@ -69,32 +66,4 @@ func (self *Slip) Exec(src string) Object {
 	self.reader.Init(src)
 	obj := self.reader.Read()
 	return obj.Eval(self.env)
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Enviroment
-
-type Enviroment struct {
-	symbols map[string]Object
-	parent  *Enviroment
-}
-
-func NewEnviroment() *Enviroment {
-	return &Enviroment{symbols: make(map[string]Object)}
-}
-
-func NewChildEnviroment(parent *Enviroment) *Enviroment {
-	return &Enviroment{symbols: make(map[string]Object), parent: parent}
-}
-
-func (self *Enviroment) Define(sym *Symbol, obj Object) {
-	self.symbols[sym.Value] = obj
-}
-
-func (self *Enviroment) Resolve(sym *Symbol) Object {
-	obj := self.symbols[sym.Value]
-	if obj == nil && self.parent != nil {
-		return self.parent.Resolve(sym)
-	}
-	return obj
 }
